+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## ????
+- Default headers: connection, accept, user-agent
+
 ## 0.0.1 : 2023-12-02
 - Basic working version